PULSEBS-218 fixed details for code working

+++ server/daos/user_dao.js
@@ -27,7 +27,7 @@ exports.createUsersTable = function() {
 //clears the lecture table
 exports.clearUserTable = function () {
     return new Promise ((resolve,reject) =>{
-        const sql = 'DELETE FROM Users';
+        const sql = `DELETE FROM Users WHERE role!= 'officer' AND role!='manager'`;
         db.run(sql,[],(err) =>{
             if(err)
                 reject(err);
